+++ src/main/java/com/gucci/user_service/follow/dto/FollowDtoRequest.java
@@ -1,7 +1,6 @@
-package com.gucci.user_service.user.dto;
+package com.gucci.user_service.follow.dto;
 
 import jakarta.validation.Valid;
-import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import lombok.Data;
