+++ src/main/java/com/gucci/user_service/follow/service/FollowServiceImpl.java
@@ -1,10 +1,10 @@
-package com.gucci.user_service.user.service;
+package com.gucci.user_service.follow.service;
 
 import com.gucci.user_service.user.config.error.NotFoundException;
-import com.gucci.user_service.user.domain.Follow;
-import com.gucci.user_service.user.domain.FollowId;
+import com.gucci.user_service.follow.domain.Follow;
+import com.gucci.user_service.follow.domain.FollowId;
 import com.gucci.user_service.user.domain.User;
-import com.gucci.user_service.user.repository.FollowRepository;
+import com.gucci.user_service.follow.repository.FollowRepository;
 import com.gucci.user_service.user.repository.UserRepository;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
@@ -12,7 +12,6 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
-import java.util.stream.Collectors;
 
 @Service
 @RequiredArgsConstructor
