+++ src/test/java/com/gucci/user_service/user/service/FollowServiceImplTest.java
@@ -1,10 +1,11 @@
 package com.gucci.user_service.user.service;
 
+import com.gucci.user_service.follow.service.FollowServiceImpl;
 import com.gucci.user_service.user.config.error.NotFoundException;
-import com.gucci.user_service.user.domain.Follow;
-import com.gucci.user_service.user.domain.FollowId;
+import com.gucci.user_service.follow.domain.Follow;
+import com.gucci.user_service.follow.domain.FollowId;
 import com.gucci.user_service.user.domain.User;
-import com.gucci.user_service.user.repository.FollowRepository;
+import com.gucci.user_service.follow.repository.FollowRepository;
 import com.gucci.user_service.user.repository.UserRepository;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
